Move interface errors to lib/interface/errors
i.e. ContractBeingDecodedHasNoNodeError and ContractAllocationFailedError

- Update usage accordingly in lib/interface/contract
- Export in lib/interface/index
